+++ app/controllers/statics_controller.rb
@@ -1,7 +1,15 @@
 class StaticsController < ApplicationController
+  ROLE_VALUES = {
+    'Tank' => 0,
+    'Healer' => 1,
+    'Melee' => 2,
+    'Ranged' => 3
+  }.freeze
+
   before_action :find_user_statics, only: %i[index]
   before_action :find_guild, only: %i[new]
   before_action :find_static_by_slug, only: %i[show edit destroy management]
+  before_action :find_static_characters, only: %i[show]
 
   def index; end
 
@@ -41,5 +49,9 @@ class StaticsController < ApplicationController
   def find_static_by_slug
     @static = Static.find_by(slug: params[:id])
     render_error('Object is not found') if @static.nil?
+  end
+
+  def find_static_characters
+    @static_characters = ActiveModelSerializers::SerializableResource.new(@static.characters.includes(:character_class, :guild, :main_roles), each_serializer: CharacterSubscriptionSerializer).as_json[:characters].sort_by { |character| [ROLE_VALUES[character[:main_role_name]['en']], character[:character_class_name]['en'], - character[:level], character[:name]] }
   end
 end